1.0.2: Changes to support chaning of php versions with vagrant provisioning and for windows users following MKL code review

+++ scripts/100-setup_apache.sh
@@ -18,8 +18,9 @@ sed -i.bak 's/Listen 80$/Listen 8090/' /etc/apache2/ports.conf
 
 # Change user and groups to vagrant
 sed -i.bak 's/www-data$/vagrant/' /etc/apache2/envvars
-
-rm /etc/apache2/sites-available/default-ssl.conf
+if [ -f /etc/apache2/sites-available/default-ssl.conf ]; then
+    rm /etc/apache2/sites-available/default-ssl.conf
+fi
 
 # Setup VHOST Script
 yes | cp -rf /vagrant/files/vhost.sh /usr/local/bin/vhost
